+++ src/app/components/common/env_variables_form.vue
@@ -12,6 +12,13 @@
                         <input class="input" type="text" placeholder="Value" v-model="env[1]">
                     </p>
                 </div>
+                <div class="field close-field">
+                    <p class="control is-expanded">
+                        <span class="icon" @click="deleteVariable(i)" :class="{'hidden-icon': i>=value.length-1}">
+                            <i class="fas fa-times"></i>
+                        </span>
+                    </p>
+                </div>
             </div>
         </div>
     </div>
@@ -36,6 +43,10 @@ module.exports = {
     methods: {
         isDuplicate(key) {
             return this.value.filter(v => v[0] === key).length >= 2;
+        },
+        deleteVariable(index) {
+            const newVar = this.value.filter((v, i) => i !== index);
+            this.$emit("input", newVar);
         }
     }
 };
@@ -49,6 +60,14 @@ module.exports = {
     .env-variable-container {
         padding-left: 16px;
         padding-right: 16px;
+    }
+
+    .close-field {
+        align-self: center;
+
+        .hidden-icon{
+            visibility: hidden;
+        }
     }
 }
 </style>